修改美客多授权

+++ erp-sdk/erp-sdk-oms-mercado/src/main/java/com/sdk/oms/mercado/dto/mercado/order/PaymentsBeanX.java
@@ -49,11 +49,11 @@ public class PaymentsBeanX {
     @JsonProperty("order_id")
     private long orderId;
     @JsonProperty("payer_id")
-    private int payerId;
+    private long payerId;
     @JsonProperty("collector")
     private BuyerBean collector;
     @JsonProperty("card_id")
-    private int cardId;
+    private long cardId;
     @JsonProperty("site_id")
     private String siteId;
     @JsonProperty("reason")
